proj2/src/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
crispy-bootstrap5==0.7
Django==4.1.3
django-crispy-forms==1.14.0
idna==3.4
kademlia==2.2.2
pyzmq==24.0.1
requests==2.28.1
rpcudp==4.0.2
sqlparse==0.4.3
u-msgpack-python==2.7.2
urllib3==1.26.13
ntplib==0.4.0
pytz==2022.6
pycryptodome==3.16.0
loguru==0.6.0
